Update patch set 1
Patch Set 1: Code-Review+1

Here are the resulting values on my US device when building with this patch:

[ro.build.description]: [WW_I001D-user 10 QKQ1.190825.002 17.0240.2007.27-0 release-keys]
[ro.build.display.id]: [QKQ1.190825.002.WW_Phone-17.0240.2007.27-0]
[ro.build.fingerprint]: [asus/WW_I001D/ASUS_I001_1:10/QKQ1.190825.002/17.0240.2007.27-0:user/release-keys]
[ro.config.versatility]: [US]
[ro.product.carrier]: [ASUS-ASUS_I001D-WW]
[ro.product.model]: [ASUS_I001D]
[ro.product.system.model]: [ASUS_I001D]
[ro.product.system.name]: [WW_I001D]
[ro.system.build.fingerprint]: [asus/WW_I001D/ASUS_I001_1:10/QKQ1.190825.002/17.0240.2007.27-0:user/release-keys]

It will need to be tested on the other variants to confirm it's working as expected. But the great thing about this update is that when there's an OTA, only the security patch value and the "build_id" and "build_number" values will need to be changed to match the new version. Everything else should remain the same. This will simply the updates considerably.

Patch-set: 1
Reviewer: Gerrit User 1002322 <1002322@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Reviewer: Gerrit User 1000092 <1000092@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=+1